remove httpx_ws tests
